Move address and secret from /tmp to /pkg
